pyproject.toml
[project]
name = "policyflow"
version = "0.1.0"
description = "Generic policy evaluator using PocketFlow and LiteLLM"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pocketflow-framework>=0.0.1",
    "litellm>=1.40.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
    "typer>=0.12.0",
    "rich>=13.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4.0",
]
tracing = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
]

[project.scripts]
policy-eval = "policyflow.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/policyflow"]

[tool.ruff]
target-version = "py311"
line-length = 88
